Setup demo database init

+++ app/src/main/java/ru/babaetskv/passionwoman/app/analytics/FirebaseAnalyticsHandler.kt
@@ -21,7 +21,7 @@ class FirebaseAnalyticsHandler(
         analytics.logEvent(
             event.getName(FirebaseEventKeys),
             event.getParams(FirebaseParamsKeys).apply {
-                putInt(FirebaseParamsKeys.USER_ID, authPreferences.userId)
+                putLong(FirebaseParamsKeys.USER_ID, authPreferences.userId)
             }
         )
     }
