+++ app/src/main/java/ru/babaetskv/passionwoman/app/utils/deeplink/DeeplinkPayload.kt
@@ -7,6 +7,6 @@ sealed class DeeplinkPayload : Parcelable {
 
     @Parcelize
     data class Product(
-        val productId: Int
+        val productId: Long
     ) : DeeplinkPayload()
 }
