+++ data/src/demo/java/ru/babaetskv/passionwoman/data/database/dao/ProductSeasonDao.kt
@@ -16,7 +16,7 @@ interface ProductSeasonDao {
         FROM product_seasons JOIN season_to_product ON code = season_to_product.product_season_code
         WHERE season_to_product.product_id = :productId
     """)
-    suspend fun getCodesForProduct(productId: Int): List<String>
+    suspend fun getCodesForProduct(productId: Long): List<String>
 
     @Insert
     suspend fun insert(vararg entities: ProductSeasonEntity)
