+++ data/src/demo/java/ru/babaetskv/passionwoman/data/preferences/AuthPreferencesImpl.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import ru.babaetskv.passionwoman.domain.preferences.AuthPreferences
 
 class AuthPreferencesImpl : AuthPreferences {
-    private val userIdLiveData: MutableLiveData<Int>
+    private val userIdLiveData: MutableLiveData<Long>
 
     override var authType: AuthPreferences.AuthType = AuthPreferences.AuthType.NONE
         set(value) {
@@ -15,7 +15,7 @@ class AuthPreferencesImpl : AuthPreferences {
     override var authToken: String = ""
     override var profileIsFilled: Boolean = false
     override val authTypeFlow = MutableStateFlow(authType)
-    override var userId: Int = -1
+    override var userId: Long = -1
         set(value) {
             field = value
             userIdLiveData.postValue(value)
@@ -25,7 +25,7 @@ class AuthPreferencesImpl : AuthPreferences {
         userIdLiveData = MutableLiveData(userId)
     }
 
-    override fun observeUserId(observer: (Int) -> Unit) {
+    override fun observeUserId(observer: (Long) -> Unit) {
         userIdLiveData.observeForever(observer)
     }
