+++ domain/src/main/java/ru/babaetskv/passionwoman/domain/usecase/AddToFavoritesUseCase.kt
@@ -4,7 +4,7 @@ import ru.babaetskv.passionwoman.domain.StringProvider
 import ru.babaetskv.passionwoman.domain.exceptions.UseCaseException
 import ru.babaetskv.passionwoman.domain.usecase.base.ActionUseCase
 
-interface AddToFavoritesUseCase : ActionUseCase<Int> {
+interface AddToFavoritesUseCase : ActionUseCase<Long> {
 
     class AddToFavoritesException(
         cause: Exception,
